Project diff rebuild
