+++ crates/editor/src/editor.rs
@@ -338,7 +338,6 @@ pub fn init(cx: &mut App) {
             .detach();
         }
     });
-    git::project_diff::init(cx);
 }
 
 pub struct SearchWithinRange;
