+++ crates/editor/src/git.rs
@@ -1,2 +1,1 @@
 pub mod blame;
-pub mod project_diff;
