+++ crates/git_ui/Cargo.toml
@@ -20,6 +20,8 @@ editor.workspace = true
 futures.workspace = true
 git.workspace = true
 gpui.workspace = true
+language.workspace = true
+multi_buffer.workspace = true
 menu.workspace = true
 project.workspace = true
 schemars.workspace = true
